+++ BTCPayServer/Configuration/BTCPayServerOptions.cs
@@ -72,9 +72,9 @@ namespace BTCPayServer.Configuration
                 else
                 {
                     if (conf.GetOrDefault<string>("SQLITEFILE", null) != null)
-                        throw new ConfigException("SQLITE backend support is out of support. Please migrate to Postgres by following the following instructions (https://github.com/btcpayserver/btcpayserver/blob/master/docs/db-migration.md). If you don't want to update, you can try to start this instance by using the command line argument --deprecated");
+                        throw new ConfigException("SQLITE backend support is out of support. Please migrate to Postgres by following the historical instructions (https://github.com/btcpayserver/btcpayserver/blob/v1.13.7/docs/db-migration.md). If you don't want to update, you can try to start this instance by using the command line argument --deprecated");
                     if (conf.GetOrDefault<string>("MYSQL", null) != null)
-                        throw new ConfigException("MYSQL backend support is out of support. Please migrate to Postgres by following the following instructions (https://github.com/btcpayserver/btcpayserver/blob/master/docs/db-migration.md). If you don't want to update, you can try to start this instance by using the command line argument --deprecated");
+                        throw new ConfigException("MYSQL backend support is out of support. Please migrate to Postgres by following the historical instructions (https://github.com/btcpayserver/btcpayserver/blob/v1.13.7/docs/db-migration.md). If you don't want to update, you can try to start this instance by using the command line argument --deprecated");
                 }
             }
             DockerDeployment = conf.GetOrDefault<bool>("dockerdeployment", true);